Refactor PublicEmployeeForm to streamline form submission process by utilizing StorePublicFormSubmissionAction; remove redundant attachment handling and improve error handling for duplicate submissions.
